Add std::allocator which preferes PSRAM
Adda custom allocators which will use the PSRAM of the ESP32 first, when allocating a string. With this and the usage of std::string we can use the default destructor to destory the Playlist std::vector.

TODO: check if default allocator (with malloc) does not already use the PSRAM (with regards to CONFIG_SPIRAM_USE_MALLOC)