Use dynamically allocated buffer for file upload
Allocate the double buffered memory for file upload dynamically. Use the same memory block for the whole progress.
If we fail the allocation we retry 2 times, always reducing the cunk size by half. If we either reach the retry count or the buffer falls below 256 bytes, then we just give up.

Also modify the size of the buffer according to the avaliable RAM as proposed by tueddy